ref: move init code to `bufferline.setup`
This is more in line with other Lua plugins these days which use
lazy initialization (i.e. they don't init until users run
`require'foo'.setup(my_options)`). While this plugin still automatically
initializes, it puts us in a better position to do lazy loading if we
actually wanted to do that in the future. Also, it allows us to localize
the `DEFAULT_OPTIONS` setting and the `local foo = vim.api.nvim_xyz`.
